Template: vite-cf-DO-v2-runner
Base template for Create a pickleball round robin game scheduler for individuals who will be playing doubles pickleball games against and with new players. The number of games will be determined by the number of players and the number of courts which must be determined first. If there are not enough courts, players will get byes until they can play.